requirements.txt
HEBO==0.3.6
about-time==4.2.1
accelerate==1.8.1
alive-progress==3.2.0
annotated-types==0.7.0
autograd==1.8.0
biopython==1.85
catboost==1.2.8
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
cma==3.2.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.2
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
disjoint_set==0.8.0
eval_type_backport==0.2.2
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
GPy==1.13.2
GPyOpt==1.2.6
gpytorch==1.13
grapheme==0.6.0
graphviz==0.21
hf-xet==1.1.5
hydra-core==1.3.2
omegaconf==2.3.0
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
# Editable install with no version control (InstructNA_frameworks==0.1)
-e .
jaxtyping==0.2.19
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.7
linear-operator==0.5.3
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
munkres==1.1.4
narwhals==1.47.0
networkx==3.2.1
numpy==1.24.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.3.1
paramz==0.9.6
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
protobuf==6.31.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pymoo==0.6.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==2.33.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
-e git+https://github.com/Hzzone/torch_clustering@b7fb417f7a144a8332ed9f9417e212f2a6ea7c05#egg=torch_clustering
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.0
wrapt==1.17.2
zipp==3.23.0
